+++ src/structvisualizer/animatecollections/animatemethods/arraylist/AnimateMethodSetArrayList.java
@@ -1,0 +1,20 @@
+package structvisualizer.animatecollections.animatemethods.arraylist;
+
+ /*
+ * AnimateMethodSetArrayList   4/3/16, 22:29
+ *
+ * By Kyrylo Havrylenko
+ *
+ */
+
+import structvisualizer.animatecollections.animatemethods.AnimateMethod;
+
+/**
+ * <what class do>
+ *
+ * @author Kyrylo Havrylenko
+ * @see
+ */
+public class AnimateMethodSetArrayList extends AnimateMethod {
+
+}
