+++ src/test/structvisualizer/valuefactories/TooltipValueFactoryTest.java
@@ -1,24 +1,26 @@
-package structvisualizer.valuefactories.Test;
-
- /*
- * TooltipValueFactoryTest   3/25/16, 19:00
- *
- * By Kyrylo Havrylenko
- *
- */
+package test.structvisualizer.valuefactories;
 
 import org.junit.Assert;
 import org.junit.Test;
 import structvisualizer.valuefactories.TooltipValueFactory;
 
+import static org.junit.Assert.*;
+
+
+ /*
+ * TooltipValueFactoryTest   4/3/16, 23:33
+ *
+ * By Kyrylo Havrylenko
+ *
+ */
+
 /**
- * Tests for TooltipValueFactory
+ * <what class do>
  *
  * @author Kyrylo Havrylenko
- * @see TooltipValueFactory
+ * @see
  */
 public class TooltipValueFactoryTest {
-
     @Test
     public void getStringValueShouldReturn() {
         Character c = 'a';
@@ -27,4 +29,5 @@ public class TooltipValueFactoryTest {
         String actual = "" + c + c;
         Assert.assertEquals(TooltipValueFactory.getStringValue(1), actual);
     }
+
 }
